Fixed bug in save data
when the spot market is turned off after base year, then the timeseries is smaller than the others. The assumption in the code is that all time series had the same size. This is not the case anymore with the latest feature where interventions are activated with *future.yaml after the base year in the forecast. (2019)